requirements.txt
dvc[s3]==2.35.2
pytorch-lightning==1.8.6
pydantic==1.10.2
torchmetrics==0.11.0
omegaconf==2.2.3
tensorboard==2.11.0
wandb==0.13.7
typer==0.7.0
pandas==1.5.2
ogb==1.3.5
scikit-learn==1.1.3
tabulate==0.9.0
shyaml==0.6.2
tbparse==0.0.7
scipy==1.9.3

jupyterlab==3.5.1
papermill==2.4.0
ipywidgets==8.0.3
seaborn==0.12.1
